# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/item/create": {
            "post": {
                "description": "创建失物（type=0）或招领（type=1）信息，创建即发布（status=0）；\u003cbr /\u003euser_id 取自 JWT；可携带 tag_ids 关联标签；location_id 指向不存在地点返回 20010，标签不存在返回 40001",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户创建物品信息",
                "parameters": [
                    {
                        "description": "创建物品请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/delete": {
            "post": {
                "description": "仅发布者本人可删；items 表为逻辑删除，删除后 is_deleted=1，不再出现在任何列表/详情中",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户删除自己的物品",
                "parameters": [
                    {
                        "description": "删除物品请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/list": {
            "get": {
                "description": "按类型/状态/地点/标签/关键词筛选物品，分页返回 ItemResponse 数组\u003cbr /\u003e未指定 status 时默认只返回已发布（status=0）的未删除物品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "公开分页查询物品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "类型: 0丢失 1拾到",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态: 0已发布 1已认领 2已关闭，默认0",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "地点ID",
                        "name": "location_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "标签ID",
                        "name": "tag_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标题/描述关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大100",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ItemListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/item/mine": {
            "get": {
                "description": "查询当前用户发布的物品，默认返回全部状态；筛选与分页参数同公开列表接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户查询自己的发布记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "类型: 0丢失 1拾到",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态: 0已发布 1已认领 2已关闭，默认全部",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标题/描述关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量，默认10，最大50",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ItemListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/item/update": {
            "post": {
                "description": "仅发布者本人可改；给定 id 和非空字段做增量更新，非整体替换；\u003cbr /\u003e携带 tag_ids（可为空数组）时整体替换标签关联；非本人返回 20005，物品不存在返回 20001",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户增量更新自己的物品",
                "parameters": [
                    {
                        "description": "更新物品请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}": {
            "get": {
                "description": "返回物品详情（含地点链、标签、图片），浏览量 +1；物品不存在或已删除返回 20001",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "公开获取物品详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.CommonResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ItemResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}/claim": {
            "post": {
                "description": "仅 status=0（已发布）物品可认领；不能认领自己发布的物品；\u003cbr /\u003eserver.claim_qq_required 开启时，未绑定QQ的用户返回 30006；\u003cbr /\u003e已被他人认领返回 30001，已关闭返回 30005，物品不存在返回 20001",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户认领物品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}/claim/cancel": {
            "post": {
                "description": "仅物品处于 status=1（已认领）且未关闭时可撤回；\u003cbr /\u003e认领者和发帖者双方均可撤回，撤回后物品恢复为 status=0 可再次被认领；\u003cbr /\u003e已关闭返回 30005，无认领返回 30002，无权撤回返回 30003",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "撤回认领（认领者或发帖者）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}/close": {
            "post": {
                "description": "仅发布者本人可操作；status=0（已发布）或 status=1（已认领）均可关闭；\u003cbr /\u003e关闭后 status=2 不可再认领/撤回，不发积分（若需确认他人认领并发分请用 confirm 接口）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "发帖者关闭自己的帖子",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}/confirm": {
            "post": {
                "description": "仅发布者本人可操作，且物品需处于 status=1（已认领）；\u003cbr /\u003e确认后物品关闭（status=2），并按 server.claim_credit 给拾到者加积分：\u003cbr /\u003e拾物帖(type=1)给发帖者，失物帖(type=0)给认领者；\u003cbr /\u003e无认领可确认返回 30002，已关闭返回 30005，非本人返回 20005",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "发帖者确认认领并关闭物品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/item/{itemID}/images": {
            "post": {
                "description": "仅发布者本人可操作；整体替换该物品的图片列表，最多3张，sort_order 取值1-3且不可重复；\u003cbr /\u003e传空数组 images=[] 清空图片；超过3张返回 20013，参数非法返回 20014",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "item"
                ],
                "summary": "登录用户覆盖式设置自己物品的图片",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "物品ID",
                        "name": "itemID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "覆盖式设置图片请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SetItemImagesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateItemRequest": {
            "type": "object",
            "required": [
                "description",
                "lost_found_time",
                "title",
                "type"
            ],
            "properties": {
                "contact": {
                    "type": "string"
                },
                "credit_reward": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "location_detail": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "lost_found_time": {
                    "type": "string"
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "model.DeleteItemRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.ItemImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "item_id": {
                    "type": "integer"
                },
                "sort_order": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.ItemListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ItemResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.ItemResponse": {
            "type": "object",
            "properties": {
                "claim_time": {
                    "type": "string"
                },
                "claim_user_id": {
                    "type": "integer"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "credit_reward": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ItemImage"
                    }
                },
                "location_detail": {
                    "type": "string"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Location"
                    }
                },
                "lost_found_time": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "model.SetItemImagesRequest": {
            "type": "object",
            "required": [
                "images"
            ],
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ItemImageInput"
                    }
                }
            }
        },
...
        "model.UpdateItemRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "contact": {
                    "type": "string"
                },
                "credit_reward": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location_detail": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "lost_found_time": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tag_ids": {
                    "description": "TagIDs 非 nil 时整体替换物品标签关联",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                10001,
                10002,
                10003,
                10004,
                10005,
                10006,
                10007,
                10008,
                10009,
                10010,
                10011,
                10012,
                10013,
                10014,
                20001,
                20005,
                20006,
                20010,
                20013,
                20014,
                30001,
                30002,
                30003,
                30004,
                30005,
                30006,
                40001,
                40002,
                40004,
                40005,
                80001,
                80002,
                80004,
                80005,
                -1
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeParamError",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeServerError",
                "CodeDatabaseError",
                "CodeUnknownError",
                "CodeChenSongError",
                "CodeUploadFileTooLarge",
                "CodeUploadFileTypeInvalid",
                "CodeUploadFailed",
                "CodeUserOrPasswordError",
                "CodeUsernameOccupied",
                "CodeFormInvalid",
                "CodeInvalidSigningMethod",
                "CodeTokenBanned",
                "CodeUserNotFoundOrBanned",
                "CodeQQSessionAlreadyExists",
                "CodeQQCodeError",
                "CodeQQUserNotInGroup",
                "CodeQQNumberError",
                "CodeQQTooManyRequests",
                "CodeQQSessionNotExist",
                "CodeQQAlreadyRegistered",
                "CodeCreditNotEnough",
                "CodeItemNotFound",
                "CodeItemNoPermission",
                "CodeItemTypeInvalid",
                "CodeItemLocationInvalid",
                "CodeItemImageTooMany",
                "CodeItemImageInvalid",
                "CodeItemAlreadyClaimed",
                "CodeClaimNotFound",
                "CodeClaimNoPermission",
                "CodeClaimSelfItem",
                "CodeItemClosed",
                "CodeClaimQQRequired",
                "CodeTagNotFound",
                "CodeTagDuplicate",
                "CodeTagNameInvalid",
                "CodeTagInUse",
                "CodeLocationNotFound",
                "CodeLocationDuplicate",
                "CodeLocationInUse",
                "CodeLocationHasChildren",
                "CodeTest"
            ]
        },
        "response.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.Code"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}